+++ src/test/java/com/lmt/lib/bldt/UpdateResultTest.java
@@ -1,0 +1,12 @@
+package com.lmt.lib.bldt;
+
+import org.junit.jupiter.api.Test;
+
+public class UpdateResultTest {
+	// このクラスは ContentDatabase#update(HttpClient, Duration, UpdateProgress, Map<String, UpdateResult>) の
+	// テストにてクラス全体のテストを網羅しているため、実装しない。
+	@Test
+	public void test() {
+		// nop
+	}
+}