+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2025-08-04
+### Added
+- Solar難易度表をサポートしました。(Presets.SOLAR)
+- Supernova難易度表をサポートしました。(Presets.SUPERNOVA)
+- ContentDatabase#update() で、難易度表更新途中で例外がスローされても全難易度表の更新が実行されるオーバーロードメソッドを追加しました。
+
 ## [0.1.1] - 2025-06-20
 ### Added
 - 癖譜面ライブラリーを標準サポートしました。(Presets.UNIQUE)